requirements.txt
django
django-allauth
qrcode[pil]

# dev only, normally in an other requirements file
django-debug-toolbar
